// frontend/packages/blockly/src/types.ts
export interface BlocklySerialization {
  blocks?: {
    blocks?: Blockly.serialization.blocks.State[];
  };
}